% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_image.R
\name{add_image}
\alias{add_image}
\title{Add a JPG/PNG to a plot}
\usage{
add_image(
  image,
  file,
  x,
  y,
  adj = c(0.5, 0.5),
  size = NULL,
  col = NA,
  alpha = NULL,
  angle = 0,
  fill = NA,
  border = NA,
  padding = NULL,
  add = TRUE,
  ...
)
}
\arguments{
\item{image}{...}

\item{file}{...}

\item{x}{...}

\item{y}{...}

\item{adj}{...}

\item{size}{...}

\item{col}{...}

\item{alpha}{...}

\item{angle}{...}

\item{fill}{...}

\item{border}{...}

\item{padding}{...}

\item{add}{...}

\item{...}{...}
}
\description{
Add a JPG/PNG to a plot
}
