% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{build_plan}
\alias{build_plan}
\title{Build tibble containing the plan}
\usage{
build_plan(
  default_start,
  default_end,
  default_days,
  wday_names,
  default_holidays,
  default_topics
)
}
\value{

}
\description{
Build tibble containing the plan
}
\author{
Fernando Naufel
}
