% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-package.R
\docType{package}
\name{calendar-package}
\alias{calendar}
\alias{calendar-package}
\title{calendar: Create a Monthly Calendar}
\description{
Plots and exports a monthly calendar as 'pdf'. Holidays (France only) are identified. User can easily add events (up to four by day) and choose between of a week of five days (working week) or a full week (weekend included).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/calendar}
  \item Report bugs at \url{https://github.com/frbcesab/calendar/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
