% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_errors.R
\name{error_modal}
\alias{error_modal}
\title{Shows modal window with error message}
\usage{
error_modal(msg)
}
\arguments{
\item{msg}{}
}
\value{
None
}
\description{
Shows modal window with error message
}
\author{
Fernando Naufel
}
