% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{expand_holidays}
\alias{expand_holidays}
\title{expand_holidays}
\usage{
expand_holidays(df)
}
\value{
Tibble with columns ???
}
\description{
Transform holidays df: expand multiple-day holidays to several rows
}
