% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_errors.R
\name{feedback_modal}
\alias{feedback_modal}
\title{Shows modal window of specified type}
\usage{
feedback_modal(msg, type, session = getDefaultReactiveDomain())
}
\arguments{
\item{msg}{}

\item{type}{}
}
\value{
None
}
\description{
Shows modal window of specified type
}
\author{
Fernando Naufel
}
