% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_events.R
\name{filter_events}
\alias{filter_events}
\title{Filter calendar events to match the extent on the calendar}
\usage{
filter_events(
  data,
  year = format(Sys.Date(), "\%Y"),
  month = format(Sys.Date(), "\%m"),
  format = "\%Y-\%m-\%d",
  weekend = FALSE
)
}
\arguments{
\item{data}{a \code{data.frame} with at least the following columns: \code{from}, the
start of the event, \code{to}, the end of the event, \code{name}, the name of the
event, and \code{category}, the category of the event (used for different
colors).}

\item{year}{either an \code{integer} or a \code{character} of length 1. Must have 4
characters (e.g. '2024' and not '24'). Default is the current year.}

\item{month}{either an \code{integer} or a \code{character} of length 1. Must have 1
or 2 characters (e.g. '01' or '1'). Default is the current month.}

\item{format}{a \code{character} of length 1. Used to specify the format of the
date. Default is \code{"\%Y-\%m-\%d"} (i.e. 2024-12-25).}

\item{weekend}{a \code{logical}. If \code{TRUE} keeps Saturdays and Sundays. Default is
\code{FALSE}.}
}
\value{
A \code{data.frame}, same as the input but with only events matching the
extent on the calendar.
}
\description{
Filter calendar events to match the extent on the calendar
}
\examples{
## No example ---
}
