% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_holidays.R
\name{filter_holidays}
\alias{filter_holidays}
\title{Filter holidays by category}
\usage{
filter_holidays(data, types)
}
\arguments{
\item{data}{a \code{data.frame} with at least the following column: \code{type},
the category of the holiday. Typically, the output of \code{get_holidays()}.}

\item{types}{a \code{character} with the types of holidays to keep.
See \code{get_holiday_types()}.}
}
\value{
A \code{data.frame}, same as the input but with only holidays matching the
\code{type} category.
}
\description{
Filter holidays by category
}
\examples{
\dontrun{
## Get holidays for United Kingdom in 2024 ----
holidays <- get_holidays(country = "UK", year = 2024, month = 4)

## Get types of holidays ----
get_holiday_types(holidays)

## Filter holidays ----
filter_holidays(holidays, types = "Common Local Holiday")
}
}
