% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calendar.R
\name{get_calendar}
\alias{get_calendar}
\title{Get the calendar for a given month and year}
\usage{
get_calendar(
  year = format(Sys.Date(), "\%Y"),
  month = format(Sys.Date(), "\%m")
)
}
\arguments{
\item{year}{either an \code{integer} or a \code{character} of length 1. Must have 4
characters (e.g. '2024' and not '24'). Default is the current year.}

\item{month}{either an \code{integer} or a \code{character} of length 1. Must have 1
or 2 characters (e.g. '01' or '1'). Default is the current month.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{date}: the date of the day (\code{YYYY-MM-DD}),
\item \code{year}: the year of the day (\code{integer}),
\item \code{month}: the month of the day (\code{integer}),
\item \code{day}: the day (\code{integer}),
\item \code{week}: the number of week (\code{integer}),
\item \code{weekday}: the name of the day of the week (\code{character}),
\item \code{month_name}: the name of month (\code{character}),
\item \code{x}: position on the x-axis (used by \code{plot_calendar()}),
\item \code{y}: position on the y-axis (used by \code{plot_calendar()}).
}
}
\description{
Get the calendar for a given month and year
}
\examples{
## Calendar for the current month ----
head(get_calendar())

## Calendar for December (current year) ----
head(get_calendar(month = 12))

## Calendar for April (current year) ----
head(get_calendar(month = 4))

## Calendar for January 1970 ----
head(get_calendar(year = 1970, month = 1))
}
