% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_holiday_types.R
\name{get_holiday_types}
\alias{get_holiday_types}
\title{Get holidays category}
\usage{
get_holiday_types(data)
}
\arguments{
\item{data}{a \code{data.frame} with the following column: \code{type},
the category of the holiday. Typically, the output of \code{get_holidays()}.}
}
\value{
A \code{character} with the categories of holidays.
}
\description{
Get holidays category
}
\examples{
\dontrun{
## Get holidays for United Kingdom in 2024 ----
holidays <- get_holidays(country = "UK", year = 2024, month = 4)

## Get types of holidays ----
get_holiday_types(holidays)
}
}
