% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_month_name.R
\name{get_month_name}
\alias{get_month_name}
\title{Get the name of a month}
\usage{
get_month_name(month, lang = NULL)
}
\arguments{
\item{month}{an \code{integer} of length 1. The month to convert in letters.}

\item{lang}{a \code{character} of length 1. Used to change the default locale
(i.e. the language). Default is \code{NULL} (i.e. use the current locale).
See examples below. Depending on the OS and the locale, the output can be
weird.}
}
\value{
A \code{character} of length 1.
}
\description{
Get the name of a month
}
\examples{
## Default ----
get_month_name(month = 4)

## Get Spanish name ----
get_month_name(month = 4, lang = "es_SP")

## Get Finnish name ----
get_month_name(month = 4, lang = "fi_FI")
}
