% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moon_phases.R
\name{get_moon_phases}
\alias{get_moon_phases}
\title{Get moon phases for a given year}
\usage{
get_moon_phases(year)
}
\arguments{
\item{year}{either an \code{integer} or a \code{character} of length 1. Must have 4
characters (e.g. '2024' and not '24'). Default is the current year.}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{new_moon}: the date of new moons (\code{YYYY-MM-DD}),
\item \code{full_moon}: the date of full moons (\code{YYYY-MM-DD}).
}
}
\description{
Scraps the site \url{https://www.timeanddate.com} to retrieve moon phases data.
}
\examples{
\dontrun{
## Get moon phases for 2024 ----
get_moon_phases(2024)
}
}
