% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekday_name.R
\name{get_weekday_name}
\alias{get_weekday_name}
\title{Get the weekday name of a date}
\usage{
get_weekday_name(date, format = "\%Y-\%m-\%d", lang = NULL)
}
\arguments{
\item{date}{either a \code{character} or a \code{Date} of length 1. The date to extract
the weekday name. See examples below.}

\item{format}{a \code{character} of length 1. Used to specify the format of the
date. Default is \code{"\%Y-\%m-\%d"} (i.e. 2024-12-25). See examples below.}

\item{lang}{a \code{character} of length 1. Used to change the default locale
(i.e. the language). Default is \code{NULL} (i.e. use the current locale).
See examples below. Depending on the OS and the locale, the output can be
weird.}
}
\value{
A \code{character} of length 1.
}
\description{
Get the weekday name of a date
}
\examples{
## Default ----
get_weekday_name("2024-04-01")

## Change date format ----
get_weekday_name("01/04/2024", format = "\%d/\%m/\%Y")

## Get Spanish name ----
get_weekday_name("2024-04-01", lang = "es_SP")

## Get Finnish name ----
get_weekday_name("2024-04-01", lang = "fi_FI")
}
