% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_load_holidays.R
\name{load_holidays}
\alias{load_holidays}
\title{Create holidays data frame from multiline string}
\usage{
load_holidays(text)
}
\arguments{
\item{text}{String}
}
\value{
A tibble with columns `name` (chr), `from` (date), `to` (date)
}
\description{
Create holidays data frame from multiline string
}
\author{
Fernando Naufel
}
