% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_calendar.R
\name{monthly_calendar}
\alias{monthly_calendar}
\title{Export a monthly calendar as pdf}
\usage{
monthly_calendar(
  year = format(Sys.Date(), "\%Y"),
  month = format(Sys.Date(), "\%m"),
  path = getwd(),
  filename = NULL,
  title = NULL,
  events = NULL,
  weekend = TRUE,
  palette = "#333333",
  lang = NULL,
  holidays = NULL,
  moon = FALSE,
  specials = NULL
)
}
\arguments{
\item{year}{either an \code{integer} or a \code{character} of length 1. Must have 4
characters (e.g. '2024' and not '24'). Default is the current year.}

\item{month}{either an \code{integer} or a \code{character} of length 1. Must have 1
or 2 characters (e.g. '01' or '1'). Default is the current month.}

\item{path}{a \code{character} of length 1. The directory to save the \code{pdf} file.
Must exist. Default is the current directory.}

\item{filename}{a \code{character} of length 1. The name of the \code{pdf} file.
Default is \code{calendar-YYYY-MM.pdf} (e.g. \verb{calendar-2024-04.pdf}).}

\item{title}{a \code{character} of length 1. The title of the calendar. Default is
\verb{Month YYYY} (e.g. \verb{April 2024}).}

\item{events}{an optional \code{data.frame} with the following columns: \code{name},
the name of the event, \code{from}, the starting date of the event, \code{to}, the
ending date of the event, and \code{category}, the category of the event.}

\item{weekend}{a \code{logical}. If \code{TRUE} (default) keeps Saturdays and Sundays.}

\item{palette}{a \code{character} vector of colors for the events. If only one
color is provided (default), all events will have this color. If
several colors are provided, the number of colors in \code{palette} must be
equal to the total number of event categories (duplicated colors are
accepted). Moreover, the \code{palette} argument must be a named vector, where
names match event categories. For example, let's that the \code{events} object
contains two categories (\code{cat_a} and \code{cat_b}), the \code{palette} argument must
be equal to \code{palette = c("cat_a" = "black", "cat_b" = "red")}.}

\item{lang}{a \code{character} of length 1. Used to change the default locale
(i.e. the language). Default is \code{NULL} (i.e. use the current locale).
Depending on the OS and the locale, the output can be weird.}

\item{holidays}{an optional \code{data.frame} with the following columns: \code{date},
the date of the holiday, and \code{name}, the name of the holiday. Typically,
the output of \code{get_holidays()}. Default is \code{NULL}.}

\item{moon}{a \code{logical}. If \code{TRUE} adds new/full moon glyph.
Default is \code{FALSE}.}

\item{specials}{an optional \code{data.frame} with the following columns: \code{date},
the date of the special day, and \code{name}, the name of the special day. These
special days will be added at the top of the day box. Default is \code{NULL}.}
}
\value{
No return value. The calendar will exported as a \code{pdf} file in
\code{path}.
}
\description{
Export a monthly calendar as pdf
}
\examples{
\dontrun{
## Add an example ----
}
}
