% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_days.R
\name{number_of_days}
\alias{number_of_days}
\title{Return the number of days in a month for a given year}
\usage{
number_of_days(date, format = "\%Y-\%m-\%d")
}
\arguments{
\item{date}{either a \code{character} or a \code{Date} of length 1. See examples below.}

\item{format}{a \code{character} of length 1. Used to specify the format of the
date. Default is \code{"\%Y-\%m-\%d"} (i.e. 2024-12-25). See examples below.}
}
\value{
An \code{integer} corresponding to the number of days in the month for
the specified year.
}
\description{
Return the number of days in a month for a given year
}
\examples{
## Default ----
number_of_days("2024-02-01")

## Day doesn't matter ----
number_of_days("2024-02-23")

## Year matters ----
number_of_days("2023-02-23")

## Change the format ----
number_of_days("2024/02/23", format = "\%Y/\%m/\%d")

## French format ----
number_of_days("23/02/2024", format = "\%d/\%m/\%Y")
}
