% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_errors.R
\name{warning_modal}
\alias{warning_modal}
\title{Shows modal window with warning message}
\usage{
warning_modal(msg, session)
}
\arguments{
\item{msg}{}
}
\value{
None
}
\description{
Shows modal window with warning message
}
\author{
Fernando Naufel
}
