% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_holidays.R
\name{get_holidays}
\alias{get_holidays}
\title{Get holidays for a given country and a given year}
\usage{
get_holidays(country, year)
}
\arguments{
\item{country}{a \code{character} of length 1. The name of the country
(e.g. \code{'France'}).}

\item{year}{either an \code{integer} or a \code{character} of length 1. Must have 4
characters (e.g. '2024' and not '24').}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{date}: the date of the holiday (\code{YYYY-MM-DD}),
\item \code{name}: the name of the holiday (\code{character}).
}
}
\description{
Scraps the site \url{https://www.timeanddate.com} to retrieve holidays data.
}
\examples{
## Get holidays for France in 2024 ----
get_holidays("France", 2024)
}
