% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-package.R
\docType{package}
\name{planner-package}
\alias{planner}
\alias{planner-package}
\title{planner: Create Ready to Print Calendars}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates annual, monthly, and weekly calendars as PDF files in the A4 paper format. Various information can be added: events, holidays, special days, moon phases, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/planner}
  \item Report bugs at \url{https://github.com/frbcesab/planner/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
